Fix babel-plugin-transform-es2015-modules-commonjs tests
