+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-from-7/actual.js
@@ -1,0 +1,1 @@
+export {default as foo} from "foo";
