+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-from-7/expected.js
@@ -1,0 +1,14 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _foo = require("foo");
+
+Object.defineProperty(exports, "foo", {
+  enumerable: true,
+  get: function () {
+    return babelHelpers.interopRequireDefault(_foo).default;
+  }
+});
