+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-named-2/expected.js
@@ -4,9 +4,4 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 exports.foo = foo;
-exports.foo = foo;
-exports.bar = bar;
-exports.bar = foo;
-exports.default = foo;
-exports.default = foo;
 exports.bar = bar;
