+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-named-4/actual.js
@@ -1,0 +1,1 @@
+export {foo as default};
