+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-named-4/expected.js
@@ -1,0 +1,6 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = foo;
