+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-named-5/expected.js
@@ -1,0 +1,7 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = foo;
+exports.bar = bar;
