+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/export-named/expected.js
@@ -1,0 +1,6 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.foo = foo;
