--- packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/exports-from/actual.js
@@ -1,7 +1,0 @@
-export * from "foo";
-export {foo} from "foo";
-export {foo, bar} from "foo";
-export {foo as bar} from "foo";
-export {foo as default} from "foo";
-export {foo as default, bar} from "foo";
-export {default as foo} from "foo";
