+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/exports-variable/expected.js
@@ -3,15 +3,15 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.foo7 = foo7;
-var foo = exports.foo = 1;
-var foo = exports.foo = 1,
-    bar = exports.bar = 2;
-var foo2 = exports.foo2 = function () {};
-var foo3 = exports.foo3 = undefined;
-let foo4 = exports.foo4 = 2;
-let foo5 = exports.foo5 = undefined;
-const foo6 = exports.foo6 = 3;
-function foo7() {}
-class foo8 {}
 exports.foo8 = foo8;
+var foo = exports.foo = 1;
+var foo2 = exports.foo2 = 1,
+    bar = exports.bar = 2;
+var foo3 = exports.foo3 = function () {};
+var foo4 = exports.foo4 = undefined;
+let foo5 = exports.foo5 = 2;
+let foo6 = exports.foo6 = undefined;
+const foo7 = exports.foo7 = 3;
+function foo8() {}
+class foo9 {}
+exports.foo9 = foo9;
