+++ packages/babel-plugin-transform-es2015-modules-commonjs/test/fixtures/interop/overview/actual.js
@@ -7,7 +7,7 @@ import {bar} from "foo4";
 import {foo as bar2} from "foo5";
 
 export {test};
-export var test = 5;
+export var test2 = 5;
 
 bar;
 bar2;
